+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# Version 1.1_r1:
+- Added Check for updates button;
+- Added Changelog;
+- Added Permissions layout;
+- Fixed some icons;
+- Added some button bindings.
+
 # Version 1.0_r1:
 - Initial stable version released.